messing with git bash
