--- making-crud/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>MY APP</title>
-</head>
-<body>
-    <h1>May Node and Express be with you. </h1>
-
-    <form action="/quotes" method="post">
-        <input type="text" name="name" placeholder="name">
-        <input type="text" name="quote" placeholder="quote">
-        <button type="submit">Submit</button>
-    </form>
-</body>
-</html>
