+++ views/index.ejs
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>MY APP</title>
+</head>
+
+<body>
+    <h1>May Node and Express be with you. </h1>
+    <form action="/quotes" method="post">
+        <input type="text" name="name" placeholder="name">
+        <input type="text" name="quote" placeholder="quote">
+        <button type="submit">Submit</button>
+    </form>
+    <div>
+        <h2>Darth Vadar invades!</h2>
+        <p>
+            Replace first Yoda's quote with a quote written by Darth Vadar
+        </p>
+        <button id="update-button">Replace Yoda's quote</button>
+    </div>
+    <div>
+        <h2>Remove Darth Vadar!</h2>
+        <p>
+            Delete one Darth Vadar's quote. Does nothing if there are no more Darth Vadar's quote
+        </p>
+        <button id="delete-button">Delete Darth Vadar's quote</button>
+    </div>
+    <div id="message"></div>
+    <h2>Quotes</h2>
+    <ul class="quotes">
+        <!--Loop thought quotes-->
+        
+        <% for(var i=0; i<quotes.length; i++) {%>
+            <li class="quote">
+                <span><%= quotes[i].name %></span>: 
+                <span><%= quotes[i].quote %></span>
+            </li>
+            <% } %>
+    </ul>
+    <script src="/main.js"></script>
+</body>
+
+</html>
